remove redundant mocks
